Migrating base package to org.ideoholic
