--- src/main/java/com/howtodoinjava/app/dao/ItemRepository.java
@@ -1,10 +1,0 @@
-package com.howtodoinjava.app.dao;
-
-import com.howtodoinjava.app.model.Item;
-import org.springframework.data.repository.ListCrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ItemRepository extends ListCrudRepository<Item, Long> {
-
-}
