+++ src/main/java/org/ideoholic/mrs/dao/ItemRepository.java
@@ -1,0 +1,10 @@
+package org.ideoholic.mrs.dao;
+
+import org.ideoholic.mrs.model.Item;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ItemRepository extends JpaRepository<Item, Long> {
+
+}
